\name{BMhyd-internal}
\alias{AICc}
\alias{AdaptiveConfidenceIntervalSampling}
\alias{AdjustForDet}
\alias{AkaikeWeight}
\alias{AttachHybridsToDonor}
\alias{CalculateLikelihood}
\alias{DetPass}
\alias{GenerateValues}
\alias{GetAncestor}
\alias{GetClade}
\alias{GetMeansModified}
\alias{GetVModified}
\alias{LumpIntoClades}
\alias{PlotNetwork}
\alias{SimulateNetwork}
\alias{SimulateTipData}
\alias{se.function}
\alias{var.model.Index.function}
\alias{weight.para.value}
\title{
Internal functions
}
\description{
These functions lack further documentation at this time 
}


\author{
Dwueng-Chwuan Jhwueng <djhwueng@umail.iu.edu>, Brian O'Meara <bomeara@utk.edu>
}
