  \name{cichlid}
     \docType{data}
     \alias{cichlid}
     \alias{network}
     \alias{data}
     \title{Cichlid data from Koblmuller et al. 2007}
     \description{
       This gives the network and data for the extant species from Koblmuller et al. 2007.
     }
     \usage{cichlid}
     \format{An object in eNewick format and the tip data for the species. Note it includes extinct taxa}
     \source{Koblmuller et al. 2007}
     \references{
       Koblmuller, S., N. Duftner, K. M. Sefc, M. Aibara,M. Stipacek, M. Blanc, B. Egger, and 
C. Sturmbauer. 2007.Reticulate phylogeny of gastropod-shell-breeding cichlids from Lake Tanganyika: the result of repeated introgressive hybridization. BMC Evolutionary Biology 7:7.
     }
     \keyword{datasets}




