  \name{nicotiana}
     \docType{data}
     \alias{nicotitana}
     \title{nicotiana data}
     \description{
       The phylogeny, strcutre ofgene flow and comprative data for nicotiana.
     }
     \usage{nicotiana}
     \format{This gives the phylogeny ("phy" object), structure of gene flow ("flow") and comparative data for extant species ("data")}
     \source{Chase et al. (2003), Clarkson et al. (2005), Komori et al. (2000)}
     \references{
     Chase M.W., Knapp S., Cox A.V., Clarkson J.J., Butsko Y., Joseph J., Savolainen V., and  Parokonny A.S. 2003. Molecular systematics, GISH and the origin of hybrid taxa in 
           Nicotiana(Solanaceae). Annals of Botany 92: 107-127.


     Clarkson J.J., Lim K.Y., Kovarik A., Chase M.W., Knapp S. and Leitch A.R. 2005. Long-term 

	 genome diploidization I allopolyploid Nicotiana section Repandae(Solanaceae). New 

	 Phytologist 168:241-252.


     Komori T., Myers P.N., Yamada S., Kubo T., and Imaseki H. 2000. Comparative study of the Nicotiana species with respect to water deficit tolerance during early growth. 
           Euphytica 116:121-130.
     }
     \keyword{datasets}




