  \name{cichlid}
     \docType{data}
     \alias{cichlid}
     \alias{data}
     \title{Cichlid data from Kobmuller et al. 2007}
     \description{
       This gives the phylogeny, structure of gene flow and comparative data for extant species from Kobmuller et al. 2007.
     	}
     \usage{cichlid}
     \format{This gives the phylogeny, structure of gene flow and comparative data for extant species}
     \source{Kobmuller et al. 2007}
     \references{
       Kobmuller, S., N. Duftner, K. M. Sefc, M. Aibara,M. Stipacek, M. Blanc, B. Egger, and 
C. Sturmbauer. 2007.Reticulate phylogeny of gastropod-shell-breeding cichlids from Lake Tanganyika: the result of repeated introgressive hybridization. BMC Evolutionary Biology 7:7.
     }
     \keyword{datasets}




