\name{ContourFromAdaptiveSampling}
\alias{BMhyd}
\title{
Contour map from adaptive sampling
}
\description{
generate the countour map from the adaptive sampling 
}

\usage{
ContourFromAdaptiveSampling(sims, params.of.interest=NULL)
\arguments{
  \item{sims}{a data frame with the simulation results}
  \item{params.of.interest}{name of parameter that is of interest to be plotted}
  \item{col.non}{color of non hybrid taxa}
  }
\details{
This plot generate the countour plot for the parameter of interest
} 

\value{
 }

\author{
Brian O'Meara <omeara.brian@gmail.com>
}

\references{
Jhwueng D.C. and O'Meara B.C. 2015. \emph{Studying trait evolution in hybrid species on phylogenetic networks}. Submitted. 
}
\examples{}